// src/sdks/systemV2ApiDocs.ts
export interface OrgTree {
  disabled?: boolean;
  orgStatus?: number;
  abbr?: string;
  children?: OrgTree[];

  code?: number;

  name?: string;

  orgCode?: string;

  orgType?: '0' | '1' | '2' | '3';

  pcode?: number;
  region?: string;

  updateBy?: number;

  updateByName?: string;

  updateTime?: Date;
  level?: number;
}