// src/sdks/computation/computationV2ApiDocs.ts
export interface Computation {
  carbonEmission?: number;
  computationName?: string;

  companyId?: number;

  createBy?: number;

  createTime?: Date;

  dataPeriod?: '1' | '2' | '3';

  deleted?: boolean;

  id?: number;

  orgId?: number;

  orgName?: string;

  updateBy?: number;

  updateByName?: string;

  updateTime?: Date;

  year?: number;
  isoCategory_name?: string;
  ghgCategory_name?: string;
  sourceName?: string;
}